unify prompt format logic
