+++ mlflow/genai/judges/utils.py
@@ -1,5 +1,4 @@
 import json
-import re
 
 import mlflow
 from mlflow.entities.assessment import Feedback
@@ -20,16 +19,6 @@ def get_default_model() -> str:
         return "databricks"
     else:
         return "openai:/gpt-4.1-mini"
-
-
-def format_prompt(prompt: str, **values) -> str:
-    """Format double-curly variables in the prompt template."""
-    for key, value in values.items():
-        # Escape backslashes in the replacement string to prevent re.sub from interpreting
-        # them as escape sequences (e.g. \u being treated as Unicode escape)
-        replacement = str(value).replace("\\", "\\\\")
-        prompt = re.sub(r"\{\{\s*" + key + r"\s*\}\}", replacement, prompt)
-    return prompt
 
 
 def _sanitize_justification(justification: str) -> str:
